requirements.txt
accelerate==1.2.1
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.7.0
async-timeout==4.0.3
attrs==23.1.0
av==14.0.1
certifi==2023.7.22
charset-normalizer==3.3.0
datasets>=2.21.0  
dill==0.3.7
distro==1.9.0
exceptiongroup==1.2.2
fastchat==0.1.0
filelock==3.12.4
frozenlist==1.4.0
fsspec==2023.6.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.0
idna==3.4
jiter==0.8.2
joblib==1.4.2
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.4
multiprocess==0.70.15
numpy==1.26.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
openai==1.58.1
packaging==23.2
pandas==2.1.1
peft==0.14.0
pillow==11.0.0
psutil==5.9.6
pyarrow==18.1.0
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2023.3.post1
PyYAML==6.0.1
qwen-vl-utils==0.0.8
regex==2023.10.3
requests==2.32.3
rich==13.9.4
safe-rlhf==0.0.1.dev0
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.15.0
six==1.16.0
sniffio==1.3.1
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.1.0
tqdm==4.67.1
transformers==4.47.1
trl==0.13.0
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.0.7
xxhash==3.4.1
yarl==1.9.2